ลบ route และ model ที่ไม่ได้ใช้งาน

+++ app/routes/index.js
@@ -8,6 +8,7 @@ const http_status_codes_1 = require("http-status-codes");
 let shell = require("shelljs");
 var crypto = require('crypto');
 var fs = require('fs');
+const { execFile } = require('child_process');
 const moph_refer_1 = require("../middleware/moph-refer");
 const hisProvider = process.env.HIS_PROVIDER.toLowerCase();
 const resultText = './sent_result.txt';
@@ -264,7 +265,7 @@ const router = (fastify, {}, next) => {
     });
     fastify.get('/autosent-result', { preHandler: [fastify.authenticate] }, async (req, reply) => {
         try {
-            var contents = fs.readFileSync(resultText);
+            var contents = await fs.promises.readFile(resultText);
             reply.status(http_status_codes_1.StatusCodes.OK).send({
                 statusCode: http_status_codes_1.StatusCodes.OK,
                 is_set: process.env.NREFER_AUTO_SEND,
@@ -366,17 +367,18 @@ const router = (fastify, {}, next) => {
         return new Promise(async (resolve, reject) => {
             const pm2Name = api.PM2_NAME === '' ? '' : api.PM2_NAME;
             const pm2Instance = +api.PM2_INSTANCE > 0 ? +api.PM2_INSTANCE : 1;
+            if (pm2Name !== '' && !/^[\w-]{1,64}$/.test(pm2Name)) {
+                return reject(new Error('Invalid PM2_NAME'));
+            }
             console.log(' ====> restart PM2:', pm2Name, (0, moment_1.default)().locale('th').format('HH:mm:ss.SS'));
             await shell.exec('tsc');
             await shell.exec("find ./app -name '*.map' -type f -delete");
             await shell.exec('pm2 flush');
-            const shellExecute1 = `pm2 scale ${pm2Name} ${pm2Instance}`;
-            await shell.exec(shellExecute1, (err, r) => {
-                console.log(' ====> shellScaling', shellExecute1, r, (0, moment_1.default)().locale('th').format('HH:mm:ss.SS'));
+            execFile('pm2', ['scale', pm2Name, String(pm2Instance)], (err, stdout, stderr) => {
+                console.log(' ====> shellScaling', pm2Name, pm2Instance, stdout || stderr, (0, moment_1.default)().locale('th').format('HH:mm:ss.SS'));
             });
-            const shellExecute2 = `pm2 restart ${pm2Name}`;
-            shell.exec(shellExecute2, (err, shellCode) => {
-                console.log(' ====> shellCode', shellExecute2, shellCode, err, (0, moment_1.default)().locale('th').format('HH:mm:ss.SS'));
+            execFile('pm2', ['restart', pm2Name], (err, stdout, stderr) => {
+                console.log(' ====> shellCode', pm2Name, stdout || stderr, err, (0, moment_1.default)().locale('th').format('HH:mm:ss.SS'));
                 resolve(true);
             });
         });
